+++ feature-home3/src/main/java/com/materip/feature_home3/viewModel/PostBoardViewModel.kt
@@ -5,11 +5,11 @@ import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.materip.core_model.accompany_board.create.BoardRequestDto
-import com.materip.core_model.accompany_board.create.BoardStatus
-import com.materip.core_model.accompany_board.create.Category
-import com.materip.core_model.accompany_board.create.PreferredAge
-import com.materip.core_model.accompany_board.create.PreferredGender
-import com.materip.core_model.accompany_board.create.Region
+import com.materip.core_model.ui_model.BoardStatus
+import com.materip.core_model.ui_model.Category
+import com.materip.core_model.ui_model.PreferredAge
+import com.materip.core_model.ui_model.PreferredGender
+import com.materip.core_model.ui_model.Region
 import com.materip.core_repository.repository.home_repository.BoardRepository
 import com.materip.core_repository.repository.image_repository.ImageRepository
 import com.materip.feature_home3.intent.PostBoardIntent
